+++ torch/testing/_internal/common_methods_invocations.py
@@ -4,36 +4,6 @@ from typing import List, Tuple
 from torch.testing import make_non_contiguous
 
 
-# TODO: review porting these to make_tensor
-def index_variable(shape, max_indices, device=torch.device('cpu')):
-    if not isinstance(shape, tuple):
-        shape = (shape,)
-    index = torch.rand(*shape, dtype=torch.double, device=device).mul_(max_indices).floor_().long()
-    return index
-
-def gather_variable(shape, index_dim, max_indices, duplicate=False, device=torch.device('cpu')):
-    assert len(shape) == 2
-    assert index_dim < 2
-    batch_dim = 1 - index_dim
-    index = torch.zeros(*shape, dtype=torch.long, device=device)
-    for i in range(shape[index_dim]):
-        index.select(index_dim, i).copy_(
-            torch.randperm(max_indices, device=device)[:shape[batch_dim]])
-    if duplicate:
-        index.select(batch_dim, 0).copy_(index.select(batch_dim, 1))
-    return index
-
-def bernoulli_scalar():
-    return torch.tensor(0, dtype=torch.bool).bernoulli_()
-
-def mask_not_all_zeros(shape):
-    assert len(shape) > 0
-    while True:
-        result = torch.randn(shape).gt(0)
-        if result.sum() > 0:
-            return result
-
-
 # TODO: move all tri/tril/triu testing to tensor creation op test suite and remove
 #   these from here
 def _compare_trilu_indices(
